Calibrate the scale of a mizerReef model to match total observed biomass

This function replaces mizer's calibrateBiomass function to include unstructured resources. Given a Mizer Params object, it returns an updated MizerParams object which is rescaled with scaleReefModel() so that the total biomass in the model agrees with the total observed biomass.

Usage

calibrateReefBiomass(params)

Arguments

params

A MizerParams object

Value

A MizerParams object

Details

Biomass observations usually only include individuals above a certain size. This size should be specified in a biomass_cutoff column of the species parameter data frame. If this is missing, it is assumed that all sizes are included in the observed biomass, i.e., it includes larval biomass.

After using this function the total biomass in the model will match the total biomass, summed over all species. However the biomasses of the individual species will not match observations yet, with some species having biomasses that are too high and others too low. So after this function use the mizer function matchReefBiomasses() to match the biomasses for each group.

Examples

data(caribbean_3_model)
params <- calibrateReefBiomass(caribbean_3_model)