interaction matrix for a simple, generic Caribbean reef with 3 species groups

This data frame encodes the trophic interactions among three functional groups: predators, herbivores, and invertebrates. The interaction strengths set to 1 for all piscivores in this model. Herbivores consume algae and invertebrates consume both plankton and detritus, with the proportion of detritus increasing with size.

Usage

caribbean_3_interaction

Format

data frame

Source

PhD Thesis, FORCE dataset

Details

For full details on interaction matrix construction, see Chapter 3 of the PhD Thesis: Beese, C. (2025). "Modelling Coral Reef Futures: Exploring the role of structural complexity in sustaining ecosystem services." Victoria University of Wellington.

References

FORCE dataset. From: Dryden, C. (2017). Habitat structural complexity of Caribbean coral reefs and its relationships with fish community structure. PhD Thesis, Newcastle University.

Beese, C. (2025). PhD Thesis. "Modelling Coral Reef Futures: Exploring the role of structural complexity in sustaining ecosystem services." Victoria University of Wellington. https://doi.org/10.26686/wgtn.26421523