Competitive method refuge parameters for Karpata reef in Bonaire

This data frame contains refuge density parameters for a trait-based competitive refuge model. It includes start and end lengths for size bins and the proportion protected (prop_protect) for each bin, representing refuge density per square meter for ten 5cm-wide fish length bins (0–50 cm).

Usage

karpata_refuge

Format

data frame

Source

PhD Thesis, FORCE dataset

Details

Parameters are derived from the FORCE dataset (see: Dryden, C. (2017)) and field measurements from Bonaire. For details on refuge parameterization and model implementation, see Chapter 3 of the PhD Thesis: Beese, C. (2025). "Modelling Coral Reef Futures: Exploring the role of structural complexity in sustaining ecosystem services." Victoria University of Wellington.

References

FORCE dataset. From: Dryden, C. (2017). Habitat structural complexity of Caribbean coral reefs and its relationships with fish community structure. PhD Thesis, Newcastle University.

Beese, C. (2025). PhD Thesis. "Modelling Coral Reef Futures: Exploring the role of structural complexity in sustaining ecosystem services." Victoria University of Wellington. https://doi.org/10.26686/wgtn.26421523