mizerReef: A mizer extension package for modelling tropical coral reefs. Includes benthic complexity through the provision of predation refuge.

This is an extension package for the mizer package (https://sizespectrum.org/mizer/) that makes it easy to set up a mizer model with predation refuges, detritus, and algae, thereby allowing for more realistic modelling of coral reef fisheries.

This package was developed to support the creation and exploration of a model for coral reefs where we could compare areas based on their benthic complexity.

See also

Author

Maintainer: Chelsey M. Beese beese.c.m@gmail.com (ORCID)

Authors:

Gustav Delius gustav.delius@york.ac.uk (ORCID) [copyright holder]

Peter J. Mumby p.j.mumby@uq.edu.au (ORCID)

Alice Rogers alice.rogers@vuw.ac.nz (ORCID)