Switch to unstructured resource dynamics with carrying capacities

This function calculates new carrying capacities for algae and detritus resources and switches the resource dynamics from the standard linear dynamics to fluxes with a carrying capacity. The carrying capacity is set at two times the current steady state biomass of the resource.

Usage

setURcapacity(params, cap = 1.5, ...)

Arguments

params

a MizerParams object

cap

Numeric. Value to scale the steady state biomass by. Defaults to 1.5, setting the carrying capacity 50% higher than the current standing biomass.

...

Unused

Value

A mizer object with updated degradation parameters profiles

Details

See algae_dynamics_cc() and detritus_dynamics_cc() for additional detail.

See also