Calculate fisheries productivity for each functional group above a chosen minimum fishing size

Calculate fisheries productivity for each functional group above a chosen minimum fishing size

Usage

getProductivity(
  object,
  include_repro = FALSE,
  min_fishing_l = NULL,
  max_fishing_l = NULL,
  ...
)

Arguments

object

An object of class MizerParams or MizerSim. If given a MizerSim object, uses the growth rates at the final time of a simulation to calculate productivity. If given a MizerParams object, uses the initial growth rates.

include_repro

A boolean value that indicates whether to include energy for reproduction in productivity estimates. Defaults to false.

min_fishing_l

The minimum length (cm) of fished individuals for productivity estimates. Defaults to 7 cm.

max_fishing_l

The maximum length (cm) of fished individuals for productivity estimates. Defaults to max length.

...

Arguments passed on to mizer::get_size_range_array

min_w

Smallest weight in size range. Defaults to smallest weight in the model.

max_w

Largest weight in size range. Defaults to largest weight in the model.

min_l

Smallest length in size range. If supplied, this takes precedence over min_w.

max_l

Largest length in size range. If supplied, this takes precedence over max_w.

Value

If called with a MizerParams object, a vector with the productivity in grams/year/m^-2 for each functional group in the model. If called with a MizerSim object, an array (time x species) containing the productivity at each time step for all species.

Potential fisheries productivity

Fisheries productivity refers to the rate at which fish biomass is produced and available for harvest in a given area over a given period of time. Productivity cannot be measured in situ.

The productivity \(P_i(w)\) of functional group \(i\) is given by

\deqn{P_i(w) = \int_w^{w+dw} \left( N_i(w) + E_{R.i}(w) \right) w \, dw.}
     {P_i(w) = \int_w^{w+dw} ( N_i(w) + E_{R.i}(w)) w \, dw.}

\(N_i(w)\) is the abundance density (1/m^-2) and \(E_{R.i}\) is the energy rate available for growth after metabolism and movement have been accounted for (grams/year).

The productivity is calculated for all fish in the size range between min_fishing_length and max_fishing_lengthwhich can be the same for all functional groups or can be specified as a vector with one value for each species in the model.min_fishing_length defaults to \(7 cm\) regardless of functional group and max_fishing_length defaults to the maximum weight in the model.