Plot the vulnerability to predation of species by weight

When called with a MizerParams object the initial vulnerability is plotted. The complement of refuge.

Usage

plotVulnerable(
  object,
  species = NULL,
  all.sizes = FALSE,
  return_data = FALSE,
  ...
)

plotlyVulnerable(
  object,
  species = NULL,
  all.sizes = FALSE,
  return_data = FALSE,
  ...
)

Arguments

object

An object of class MizerParams

species

The species to be selected. Optional. By default all species are selected. A vector of species names, or a numeric vector with the species indices, or a logical vector indicating for each species whether it is to be selected (TRUE) or not.

all.sizes

If TRUE, then feeding level is plotted also for sizes outside a species' size range. Default FALSE.

return_data

A boolean value that determines whether the formatted data used for the plot is returned instead of the plot itself. Default value is FALSE.

...

unused

Value

A ggplot2 object, unless return_data = TRUE, in which case a data frame the vulnerability at each size

See also