Get all rates needed to project a mizerReef model

Calls other rate functions in sequence and collects the results in a list.

Usage

reefRates(params, n, n_pp, n_other, t = 0, effort, rates_fns, ...)

Arguments

params

A MizerParams object

n

A matrix of species abundances (species x size).

n_pp

A vector of the resource abundance by size

n_other

A list of abundances for other dynamical components of the ecosystem

t

The time for which to do the calculation (Not used by standard mizer rate functions but useful for extensions with time-dependent parameters.)

effort

The effort for each fishing gear

rates_fns

Named list of the functions to call to calculate the rates. Note that this list holds the functions themselves, not their names.

...

Unused

Value

List of rates.

Details

By default this function returns a list with the following components:

predation vulnerability from reefVulnerable()

encounter from reefEncounter()

feeding level from reefFeedingLevel()

e from mizerEReproAndGrowth()

e_repro from mizerERepro()

e_growth from mizerEGrowth()

pred_rate from mizerPredRate()

pred_mort from reefPredMort()

sen_mort from reefSenMort()

f_mort from mizerFMort()

mort from reefMort()

rdi from mizerRDI()

rdd from BevertonHoltRDD()

resource_mort from mizerResourceMort()

However you can replace any of these rate functions by your own rate function if you wish, see setRateFunction() for details.

See also