Checks unstructured resource parameters and interaction matrix

Checks unstructured resource parameters and interaction matrix

Usage

setURParams(
  params,
  UR_interaction = NULL,
  exp_alg = NULL,
  exp_det = NULL,
  scale_rho_a = NULL,
  scale_rho_d = NULL,
  algae_growth = NULL,
  prop_decomp = NULL,
  d.external = NULL
)

Arguments

params

MizerParams object

UR_interaction

Interaction matrix for unstructured resources (species x resource)

Optional parameters:

exp_alg

The allometric exponent for the consumption rate of algae. Defaults to 0.86.

exp_det

The allometric exponent for the consumption rate of detritus. Defaults to the same value used for the scaling exponent of the maximum intake rate for fish consumers.

scale_rho_a

A factor to multiply rho values by for algae encounter rate. Used in steady state tuning.

scale_rho_d

A factor to multiply rho values by for detritus encounter rate. Used in steady state tuning.

algae_growth

The initial growth rate of algae in grams/year/m^-2. This value is reset to match consumption in the [reefSteady()] function so that steady state abundances match given values.

prop_decomp

The proportion of waste material that decomposes to become part of the detritus pool.

d.external

The rate at which detritus biomass sinks from the pelagic zone and becomes part of the detritus pool in grams per year. This value is reset to make up any differences in consumption and production in the [reefSteady()] function so that steady state abundances match observed values.

Value

setUResourceParams: MizerParams object with updated unstructured resource parameters

Adding unstructured resources

 mizerReef supports two resource spectra that are not size- structured.
 Algae is consumed by herbivorous fish, while detritus is consumed by
 herbivorous fish and benthic invertebrates. This function sets the 
 interaction matrix for these resources as well as any default
 parameters necessary to structure them.

 The resource interaction matrix \eqn{\theta_{ki} modifies the
 interaction of each functional group \eqn{i} with each unstructured
 resource \eqn{k} in the model. This can be used for example to allow 
 for different diet preferences on each unstructured resource. 
 
 Note that interaction with size structured resources, such as
 plankton, is still set with the `resource_interaction` column of
 the species parameters dataframe.

Detritus consumption

 The rate at which detritivorous consumer groups encounter detrital 
 biomass \eqn{E_{i.D}(w)} is controlled by the parameter 
 \eqn{\rho_{D.i}}. It scales with the size of the consumer raised to 
 an allometric exponent \eqn{m_{det}} which is taken to be the same as 
 the scaling exponent of the maximum intake rate for fish consumers.
 
 \deqn{E_{i.D}(w)=\rho_{i.D}\, w^{m_{det}}\,B_D }{
           E_{i.D}(w)=\rho_{i.D}\, w^{m_{det}}\,B_D}
         
 The mass specific consumption rate then accounts for the preference of 
 functional group $i$ for detritus, \eqn{\theta_{i.D}} and the feeding 
 level \eqn{f_i(w)}. This gives the mass-specific detritus consumption
 rate:
 
 \deqn{c_D = \sum_i\int\rho_{i.D}\, w^{m_{det}} 
                 N_i(w) \left(1-f_i(w)\right) \theta_{i.D}\,dw}{
           c_D = \sum_i\int\rho_{i.D}\, w^{m_{det}} 
                 N_i(w) \left(1-f_i(w)\right) \theta_{i.D}\,dw}

Detritus production

 The rate \eqn{p_D} at which detritus biomass is produced by the 
 ecosystem has contributions from three sources:
 
 \deqn{p_D = p_{D.f} + p_{D.d} + p_{D.ext}}{
           p_D = p_{D.f} + p_{D.d} + p_{D.ext}}
       
 \eqn{p_{D.f}} comes from the biomass that is consumed but not 
 assimilated and is given by:
 
 \deqn{p_{D.f} = \sum_i(1-\alpha_i)\int E_i(w)\,dw}{
           p_{D.f} = \sum_i(1-\alpha_i)\int E_i(w)\,dw}
       
 \eqn{p_{D.d}} comes from the biomass of fish that die as a result of 
 external mortality. External mortality includes local deaths that lead 
 to detritus but also deaths due to predation by species that are not
 explicitly modelled, for example transient predators, mammals, or sea 
 birds. Thus, only a proportion `prop_decomp` of this material 
 decomposes to detritus. The detritus production from decomposing 
 dead organisms is given by:
 
 \deqn{p_{D.d} = \sum_i\int\mu_{seni.i}(w)N_i(w)w\,dw + 
                     \mathtt{prop\_decomp}\,
                     \sum_i\int\mu_{nat.i}(w)N_i(w)w\,dw}{
         p_{D.d} = \sum_i\int\mu_{seni.i}(w)N_i(w)w\,dw + 
                     \mathtt{prop\_decomp}\,
                     \sum_i\int\mu_{nat.i}(w)N_i(w)w\,dw}
                 
 \eqn{p_{D.ext}} is the rate at which detritus enters the system from
 unmodelled or external sources. For coral reefs, this includes detritus
 produced by sponges and coral mucous as well as waste material that 
 sinks in from the pelagic zone. This rate is a model parameter 
 independent of any other model component. It is set so that production 
 and consumption are equal for the chosen steady state abundances.

Algae consumption

 The rate at which herbivorous consumer groups encounter algal 
 biomass \eqn{E_{i.A}(w)} is controlled by the parameter 
 \eqn{\rho_{A.i}}. It scales with the size of the consumer raised to 
 an allometric exponent \eqn{m_alg} which is taken from empirical data.
 
 \deqn{E_{i.A}(w)=\rho_{i.A}\, w^{m_{alg}}\,B_A}{
           E_{i.A}(w)=\rho_{i.A}\, w^{m_{alg}}\,B_A}
         
 The mass specific consumption rate then accounts for the preference of 
 functional group $i$ for algae, \eqn{\theta_{i.A}}. This gives the 
 mass-specific algae consumption rate:
 
 \deqn{c_A = \sum_i\int\rho_{i.A}\, w^{m_{alg}} N_i(w)\theta_{i.A}\,dw}{
           c_A = \sum_i\int\rho_{i.A}\, w^{m_{alg}} N_i(w)\theta_{i.A}\,dw}