Stepped refuge profile for tuning steady states

This is a 2-dimensional array containing start and end lengths for size bins and prop_protect decreasing from 30% to 10% over the ten bins.

Usage

step_tune

Format

dataframe

Source

PhD Thesis

Details

This profile provides more protection to smaller size classes than larger ones, as would be observed on a natural reef.

These refuge parameters are intended for tuning the steady state when using the density-dependent competitive method. The tuning profile provides a constant proportion of refuges to all fish up to 50 cm in length.

When creating a model using the competitive method, you should run newReefParams() with the "binned" method and thing tuning profile.

After species biomasses and growth rates have been calibrated to match empirical observations, use the newRefuge() function to implement your competitive refuge parameters. After using newRefuge(), make sure to iterate through mizer:: matchBiomasses(), matchReefGrowth(), and reefSteady() again to regain the steady state.